backend/requirements.txt
google-adk
google-genai
python-dotenv 
reportlab
python-docx